Use signed types for all offsets

+++ include/quickbit.h
@@ -15,25 +15,25 @@ typedef uint8_t *quickbit_t;
 typedef uint8_t quickbit_index_t[QUICKBIT_INDEX_LEN];
 
 bool
-quickbit_get (const quickbit_t field, size_t bit);
+quickbit_get (const quickbit_t field, int64_t bit);
 
 bool
-quickbit_set (quickbit_t field, size_t bit, bool value);
+quickbit_set (quickbit_t field, int64_t bit, bool value);
 
 void
-quickbit_fill (const quickbit_t field, bool value, size_t start, size_t end);
+quickbit_fill (const quickbit_t field, bool value, int64_t start, int64_t end);
 
-size_t
-quickbit_index_of (const quickbit_t field, size_t field_len, bool value, size_t position, quickbit_index_t index);
+int64_t
+quickbit_index_of (const quickbit_t field, size_t field_len, bool value, int64_t position, quickbit_index_t index);
 
-size_t
-quickbit_last_index_of (const quickbit_t field, size_t field_len, bool value, size_t position, quickbit_index_t index);
+int64_t
+quickbit_last_index_of (const quickbit_t field, size_t field_len, bool value, int64_t position, quickbit_index_t index);
 
 void
 quickbit_index_init (quickbit_index_t index, const quickbit_t field, size_t field_len);
 
 bool
-quickbit_index_update (quickbit_index_t index, const quickbit_t field, size_t bit);
+quickbit_index_update (quickbit_index_t index, const quickbit_t field, int64_t bit);
 
 #ifdef __cplusplus
 }
